Create custom Ansible mode and make lsp-mode check for it
The Ansible mode on MELPA has font lock that can't be disabled and its
functionality is basically redundant when using ansible-language-server.
Therefore, we create a custom minor mode, ansible-mode, which only exists as a
way of enabling lsp-mode.

The minor mode also provides a little utility predicate function which can be
used to conditionally enable the minor mode only for YAML files that are
actually part of an Ansible project. The list of file names used for the
detection can be customized.

This is then hooked up to conditionally enable ansible-mode when yaml-mode is
enabled, based on whether the buffer is part of an Ansible project or not. The
lsp-mode function checking for the MELPA Ansible mode is overridden with one
that checks for the local ansible-mode instead. This way we still get all the
nice things from ansible-language-server but without annoying font lock, and
also no ansible-mode in regular YAML files.